# Dumb bitwise add