--- a/src/protocol/boolean/mod.rs
+++ b/src/protocol/boolean/mod.rs
@@ -8,6 +8,7 @@ mod bitwise_lt;
 mod bitwise_sum;
 mod carries;
 mod dumb_bitwise_lt;
+mod dumb_bitwise_sum;
 mod or;
 mod prefix_or;
 pub mod random_bits_generator;