--- a/src/protocol/boolean/or.rs
+++ b/src/protocol/boolean/or.rs
@@ -1,20 +1,19 @@
 use crate::error::Error;
 use crate::ff::Field;
 use crate::protocol::context::SemiHonestContext;
-use crate::protocol::{context::Context, mul::SecureMul, RecordId};
+use crate::protocol::{mul::SecureMul, RecordId};
 use crate::secret_sharing::Replicated;
 
-/// Secure XOR protocol with two inputs, `a, b ∈ {0,1} ⊆ F_p`.
-/// It computes `[a] + [b] - 2[ab]`
+/// Secure OR protocol with two inputs, `a, b ∈ {0,1} ⊆ F_p`.
+/// It computes `[a] + [b] - [ab]`
 pub async fn or<F: Field>(
     ctx: SemiHonestContext<'_, F>,
     record_id: RecordId,
     a: &Replicated<F>,
     b: &Replicated<F>,
 ) -> Result<Replicated<F>, Error> {
-    let one = Replicated::one(ctx.role());
-    let result = ctx.multiply(record_id, &(&one - a), &(&one - b)).await?;
-    Ok(one - &result)
+    let ab = ctx.multiply(record_id, a, b).await?;
+    Ok(a + b - &ab)
 }
 
 #[cfg(test)]