Our current BitwiseAdd protocol is VERY SLOW. I didn't understand why the current protocol was so complicated so I created a much simpler one and benchmarked it:

# add two Fp31
- Old: 35 multiplications
- New: 13 multiplications
63% reduction

# add two Fp32BitPrime
- Old: 1520 multiplications
- New: 94 multiplications
94% reduction (!!!)

This makes the tests run a LOT faster.

# Add every element of Fp31 to every other element of Fp31
- Old: 23.47 seconds
- New: 10.66 seconds
63% reduction

# Add two, randomly selected elements of Fp32BitPrime, looped 100x
- Old: 82.82 seconds
- New: 5 seconds 
94% reduction (!!!)