@taikiy 2022-11-28
Looks good to me. Maybe we can go ahead and replace the `BitwiseSum` in `BitDecomposition` and see everything works as expected.